+++ src/main/java/com/jeromerichard/pdfstream/Service/Interfaces/ArticleServiceInt.java
@@ -5,13 +5,14 @@ import com.jeromerichard.pdfstream.Entity.Article;
 import com.jeromerichard.pdfstream.Exception.EmptyListException;
 import com.jeromerichard.pdfstream.Exception.NotFoundException;
 
+import javax.lang.model.type.IntersectionType;
 import java.util.Set;
 
 public interface ArticleServiceInt {
     // CRUD JPA
     public Article saveArticle(ArticleDTODown article);
     public Set<Article> getAllArticles() throws EmptyListException;
-    public Article getArticleById(Long id) throws NotFoundException;
-    public Article updateArticle(Long id, ArticleDTODown article) throws NotFoundException;
-    public void deleteArticle(Long id) throws NotFoundException;
+    public Article getArticleById(Integer id) throws NotFoundException;
+    public Article updateArticle(Integer id, ArticleDTODown article) throws NotFoundException;
+    public void deleteArticle(Integer id) throws NotFoundException;
 }
